backend/requirements.txt
python-multipart
uvicorn
fastapi
Pillow
rembg